Se agrego mostrar rondas y total crupier

+++ src/componentes/PanelInfoCrupier.java
@@ -58,7 +58,7 @@ public class PanelInfoCrupier extends javax.swing.JPanel {
         this.ronda.setText("Ronda #" + String.valueOf(ronda));
     }
 
-    public void setTotalApostado(float totalApostado) {
+    public void setTotalApostado(int totalApostado) {
         this.lbl_totalApostado.setText(String.valueOf(totalApostado));
 
     }
